% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_caladapt.R
\name{query_caladapt}
\alias{query_caladapt}
\title{Query Cal-Adapt data}
\usage{
query_caladapt(var, model, scenario, coords, type = "events",
  timestep = "year", stat = "mean",
  url = "http://api.cal-adapt.org/api/series/")
}
\arguments{
\item{var}{Name of the variable}

\item{model}{Name of the climate model}

\item{scenario}{Name of the climate scenario (RCP): 'rcp45' or 'rcp85'}

\item{coords}{Coordinates of the location for which data is desired}

\item{type}{Return data ('events') or 'rasters'}

\item{timestep}{Annual ('year'), monthly, or daily data}

\item{stat}{Stat desired; defaults to 'mean'; unclear what other options are}

\item{url}{Cal-Adapt API URL; defaults to 'http://api.cal-adapt.org/api/series/'}
}
\value{
For type = 'events', returns a tibble. Untested for type = 'rasters'.
}
\description{
Query Cal-Adapt data
}
\details{
Options for variables include 'tasmax' and 'tasmin' (max and min
  temperatures) or 'pr' (precipitation). Note that precipitation is returned
  in units of inches per day (even for timestep = 'year'), so will need to
  be multiplied by 365 to generate annual totals. The full list of available
  models and additional variables are provided at
  \url{https://berkeley-gif.github.io/caladapt-docs/data-catalog.html#climate-variables}.
  But note that only annual average tasmin, tasmax, and pr have been checked
  against Cal-Adapt website outputs to date.

  Coordinates should be provided as a string in decimal degrees separated
  only by a comma: 'long,lat'
}
\examples{
query_caladapt(var = 'pr', scenario = 'rcp45', model = 'HadGEM2-ES', coords = '-122.545886,38.248392')

}
